https://www.hackthis.co.uk/levels/coding/1
This is taking a comma separated list of words and alphabetizing them. The reason that the output is to the clipboard is because the output needs to be generated and submitted in under 5 seconds. This is not the optimal way, as the code should be submitting for you, but this simplistic approach works for now.